Photon de mrd
Subo para q despues no se pierda. Pero es poco

+++ Assets/Scripts/PlayerHealth.cs
@@ -8,8 +8,8 @@ using UnityEngine.UI;
 
 public class PlayerHealth : MonoBehaviourPun, IDamageable
 {
-    [SerializeField] private int maxHealth = 100;
-    private int currentHealth;
+    [SerializeField] private float maxHealth = 100;
+    private float currentHealth;
 
     [SerializeField] private DamageFx damageFx;
 
@@ -26,7 +26,6 @@ public class PlayerHealth : MonoBehaviourPun, IDamageable
             damageFx = GetComponentInChildren<DamageFx>();
           
         }
-        // Inicializamos el slider y texto
         if (healthSlider != null)
         {
             healthSlider.maxValue = maxHealth;
@@ -39,9 +38,8 @@ public class PlayerHealth : MonoBehaviourPun, IDamageable
         }
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
-        //Si el objeto es mio aplico dano y efecto visual
         if (photonView.IsMine)
         {
             ApplyDamage(damage);
@@ -49,16 +47,14 @@ public class PlayerHealth : MonoBehaviourPun, IDamageable
             if (damageFx != null)
                 damageFx.ShowDamage();
         }
-        //pero si no es local llamo al RPC para que aplique dano al otro jugador en su propia maquina
         else
         {
             photonView.RPC("RPC_ApplyDamage", photonView.Owner, damage);
         }
     }
 
-    //Cuando un enemigo nos ataca el RPC aplica el dano en esta intancia y si somos duenos del objeto activa efecto visual
     [PunRPC]
-    void RPC_ApplyDamage(int damage)
+    void RPC_ApplyDamage(float damage)
     {
         ApplyDamage(damage);
 
@@ -69,7 +65,7 @@ public class PlayerHealth : MonoBehaviourPun, IDamageable
 
 
 
-    private void ApplyDamage(int damage)
+    private void ApplyDamage(float damage)
     {
         currentHealth -= damage;
